test for convertCollectionToXml
