+++ sitemap/package.json
@@ -4,7 +4,7 @@
   "description": "Generates sitemap files for all collections in the catalog",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
@@ -20,5 +20,8 @@
   "dependencies": {
     "request": "^2.88.2",
     "yargs": "^16.2.0"
+  },
+  "devDependencies": {
+    "jest": "^26.6.3"
   }
 }
